Good starting point.
Lots of changes: Sim now is sort of functional. Code needs to be cleaned
up a lot. Need to clearly define the interface for the Motor model,
Observer and Controller. Need to come up with a better method to plot /
log results. May be a clase on its own? Need to make simulation more
efficient. It gets really slow when using small (less than 1e-4)
timesteps. Need to really make sure why this happens.