Cleaned up plotting code.
Code for results plotting is more readable, however looks ugly. There
might be a better way to do it, but... whatever.

Logs that might not be used: Sensors, Observer, are now initialized to
np.nan. This is to make it easy to detect when there are no sensors
attached to the motor, or an observer is not used.

It might be useful to make the controller, observer and all that code
raise a ValueError when None is passed to the functions. This means that
the controller/observer is trying to use a signal that is not available.

TODO: Make PID work again!!!! For some reason it does not stabilize
velocity anymore.